requirements.txt
appdirs==1.4.4
asgiref==3.3.1
atomicwrites==1.4.0
attrs==20.3.0
autoflake==1.4
black==20.8b1
certifi==2020.12.5
cfgv==3.2.0
chardet==4.0.0
click==7.1.2
colorama==0.4.4
distlib==0.3.1
dj-config-url==0.1.1
dj-database-url==0.5.0
Django==3.1.4
django-cors-headers==3.6.0
django-dotenv==1.4.2
django-filter==2.4.0
django-versatileimagefield==2.0
djangorestframework==3.12.2
djangorestframework-serializer-extensions==2.0.1
filelock==3.0.12
hashids==1.3.1
identify==1.5.10
idna==2.10
importlib-metadata==3.3.0
iniconfig==1.1.1
Markdown==3.3.3
mypy-extensions==0.4.3
nodeenv==1.5.0
packaging==20.8
pathspec==0.8.1
Pillow==8.0.1
pluggy==0.13.1
pre-commit==2.9.3
psycopg2==2.8.6
py==1.10.0
pyflakes==2.2.0
pyparsing==2.4.7
pytest==6.2.1
pytest-django==4.1.0
python-decouple==3.3
python-magic==0.4.18
python-magic-bin==0.4.14
python-resize-image==1.1.19
pytz==2020.4
pyupgrade==2.7.4
PyYAML==5.3.1
regex==2020.11.13
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
tokenize-rt==4.0.0
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.26.2
virtualenv==20.2.2
zipp==3.4.0
